requirements-dev.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
astroid==3.3.11
attrs==25.3.0
certifi==2025.8.3
colorama==0.4.6
deprecation==2.1.0
dill==0.4.0
discord.py==2.3.2
frozenlist==1.7.0
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
multidict==6.6.3
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
postgrest==1.1.1
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pylint==3.3.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
realtime==2.7.0
six==1.17.0
sniffio==1.3.1
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.0
supafunc==0.10.1
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.1
websockets==15.0.1
yarl==1.20.1
audioop-lts>=0.2.1
pytest-asyncio>=0.23
pytest-mock>=3.14
tzdata>=2024.1